refactor: add file for ProposalAnswerSubevent
