+++ src/lib/events/proposal-start-event.ts
@@ -1,11 +1,7 @@
 // Ref: https://github.com/matrix-org/matrix-js-sdk/blob/develop/src/extensible_events_v1/PollStartEvent.ts
 
+import { PRAXIS_PROPOSAL_KIND } from '@/constants/proposal.constants';
 import {
-  PRAXIS_PROPOSAL_ANSWER_POSITION,
-  PRAXIS_PROPOSAL_KIND,
-} from '@/constants/proposal.constants';
-import {
-  ExtensibleAnyMessageEventContent,
   IPartialEvent,
   KnownPollKind,
   M_POLL_KIND_DISCLOSED,
@@ -18,47 +14,8 @@ import {
 import { NamespacedValue } from 'matrix-js-sdk/lib/NamespacedValue';
 import { InvalidEventError } from 'matrix-js-sdk/src/extensible_events_v1/InvalidEventError';
 import { MessageEvent } from 'matrix-js-sdk/src/extensible_events_v1/MessageEvent';
-import {
-  PollAnswerSubevent,
-  PollStartEvent,
-} from 'matrix-js-sdk/src/extensible_events_v1/PollStartEvent';
-
-export type ProposalAnswer = ExtensibleAnyMessageEventContent & {
-  id: string;
-  position?: string;
-};
-
-class ProposalAnswerSubevent extends PollAnswerSubevent {
-  public readonly id: string;
-  public readonly position?: string;
-
-  public constructor(wireFormat: IPartialEvent<ProposalAnswer>) {
-    super(wireFormat);
-
-    const id = wireFormat.content.id;
-    if (!id || typeof id !== 'string') {
-      throw new InvalidEventError('Answer ID must be a non-empty string');
-    }
-    this.id = id;
-
-    const position = wireFormat.content.position;
-    if (!position || typeof position !== 'string') {
-      throw new InvalidEventError('Answer position must be a non-empty string');
-    }
-    this.position = position;
-  }
-
-  public serialize(): IPartialEvent<object> {
-    return {
-      type: 'org.matrix.sdk.poll.answer',
-      content: {
-        id: this.id,
-        [PRAXIS_PROPOSAL_ANSWER_POSITION.name]: this.position,
-        ...this.serializeMMessageOnly(),
-      },
-    };
-  }
-}
+import { PollStartEvent } from 'matrix-js-sdk/src/extensible_events_v1/PollStartEvent';
+import { ProposalAnswerSubevent } from './proposal-answer-subevent';
 
 export class ProposalStartEvent extends PollStartEvent {
   public readonly answers: ProposalAnswerSubevent[];
